tests/wpt/tests/svg/styling/vector-effect-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>vector-effect test: parsing vector-effect with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/SVG2/coords.html#VectorEffects">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("vector-effect", "none");
test_valid_value("vector-effect", "non-scaling-stroke");
test_valid_value("vector-effect", "non-scaling-stroke viewport", "non-scaling-stroke");
test_valid_value("vector-effect", "non-scaling-stroke screen");
</script>
</body>
</html>
